Move examples into examples/ dir
